test: add event_validators for Aragon Repo events
New validator can check aragon app version upgrade event.
