+++ tests/event_validators/payout.py
@@ -3,29 +3,13 @@
 from typing import NamedTuple
 
 from brownie.network.event import EventDict
+from .common import validate_events_chain
 
 class Payout(NamedTuple):
     token_addr: str
     from_addr: str
     to_addr: str
     amount: int
-
-# Check that all tx_events contained in reference_events (with proper ordering and occurrences count).
-# Allow some of the reference_events skipped in the tx_events. 
-#
-# Examples:
-#
-# tx_events: ['A', 'B', 'C'], reference_events: ['A', 'B', 'C'] => valid
-# tx_events: ['A', 'B', 'D'], reference_events: ['A', 'B', 'C', 'D'] => valid
-#
-# tx_events: ['A', 'B', 'D'], reference_events: ['A', 'B', 'C'] => invalid // extra 'D' event
-# tx_events: ['A', 'B', 'A', 'B'], reference_events: ['A', 'B'] => invalid // duplicated 'A', 'B' events chain
-# tx_events: ['A', 'C', 'B'], reference_events: ['A', 'B', 'C'] => invalid // wrong order
-def validate_events_chain (tx_events: [str], reference_events: [str]):
-    for ev in tx_events:
-        idx = next((reference_events.index(e) for e in reference_events if e == ev), len(reference_events))
-        assert idx != len(reference_events), f"{ev} not found in the remaining {reference_events} events chain"
-        reference_events=reference_events[idx+1:]
 
 def validate_payout_event (event: EventDict, p: Payout):
     _ldo_events_chain = ['LogScriptCall', 'NewPeriod', 'NewTransaction', 'Transfer', 'VaultTransfer']
